requirements.txt
fastapi
uvicorn
google-generativeai==0.8.5